Script with python code and some slides were added.
The python script contains the solution for the problem of findind a
two-digit number which satisfies that removing the first digit it
becames 7 times smaller. Also contain the problem of finding a number
which satisfies that removing the first digit it becames 57 times
smaller. The slides contain the topic of existance proofs.